#workus/60 근무스케줄 삭제서비스 응답 수정

+++ src/main/java/com/workus/workus/attend/schedule/application/service/DeleteWorkScheduleService.java
@@ -1,9 +1,14 @@
 package com.workus.workus.attend.schedule.application.service;
 
+import com.workus.workus.attend.schedule.domain.core.WorkSchedule;
 import com.workus.workus.attend.schedule.domain.repository.WorkScheduleRepository;
+import com.workus.workus.attend.schedule.domain.violation.WorkScheduleRuleViolation;
+import com.workus.workus.common.result.Result;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -11,7 +16,13 @@ import org.springframework.stereotype.Service;
 public class DeleteWorkScheduleService {
     private final WorkScheduleRepository repository;
 
-    public void deleteWorkSchedule(Long workScheduleId){
-        repository.deleteById(workScheduleId);
+    public Result<Void, WorkScheduleRuleViolation.Delete> deleteWorkSchedule(Long workScheduleId){
+        Optional<WorkSchedule> schedule = repository.findById(workScheduleId);
+        if(schedule.isPresent()){
+            repository.deleteById(workScheduleId);
+            return Result.success(null);
+        }else{
+            return Result.failure(new WorkScheduleRuleViolation.ScheduleNotFound());
+        }
     }
 }
